ms-python/requirements.txt
fastapi
uvicorn
requests
pydantic
python-dotenv
tensorflow
nltk
pytest
pip-tools
pip-audit
httpx
